requirements.txt
python-dotenv
pandas 
numpy
setuptools
pymongo
certifi
##mongodb@3.7

##-e .
